Renamed Resource/* to ResourceModel/*
